When getting the serial port buffer size, use a vpr::Uint16 to allow for larger buffer sizes.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@4940 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ modules/vapor/vpr/md/POSIX/IO/Port/SerialPortImplTermios.cpp
@@ -155,12 +155,12 @@ SerialPortImpTermios::setUpdateAction (SerialTypes::UpdateActionOption action)
 // Query the serial port for the maximum buffer size.
 // ----------------------------------------------------------------------------
 Status
-SerialPortImpTermios::getBufferSize (Uint8& size) {
+SerialPortImpTermios::getBufferSize (Uint16& size) {
     Status retval;
     struct termios term;
 
     if ( (retval = getAttrs(&term)).success() ) {
-        size = term.c_cc[VMIN];
+        size = (Uint16) term.c_cc[VMIN];
     }
 
     return retval;
